requirements.txt
# This is the "requirements.txt" file for the weather_delay app

requests
python-dotenv
pandas
pytest
